+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [yuyo.backoff.Backoff][] now increments the internal counter regardless of whether
   [yuyo.backoff.Backoff.set_next_backoff][] has been called.
+- [yuyo.backoff.Backoff][] now iterates over the retry counter ([int][]), starting at 0.
+  Rather than just [None][].
 
 ### Fixed
 - [yuyo.backoff.Backoff.backoff][] now respects the max retires config and finished flag.